Refactor module post

+++ modules/datacenter/post/entity.go
@@ -12,24 +12,24 @@ var (
 
 // Post represents a blog post entity
 type Post struct {
-	ID        int64     `json:"id"`         // ID
+	ID        uint      `json:"id"`         // ID
 	Title     string    `json:"title"`      // 标题
 	Content   string    `json:"content"`    // 内容
 	Slug      string    `json:"slug"`       // 别名
 	Status    string    `json:"status"`     // 状态：draft, published
-	AuthorID  int64     `json:"author_id"`  // 作者ID
+	AuthorID  uint      `json:"author_id"`  // 作者ID
 	CreatedAt time.Time `json:"created_at"` // 创建时间
 	UpdatedAt time.Time `json:"updated_at"` // 更新时间
 }
 
 // PostResponse represents the response structure for post data
 type PostResponse struct {
-	ID        int64  `json:"id"`         // ID
+	ID        uint   `json:"id"`         // ID
 	Title     string `json:"title"`      // 标题
 	Content   string `json:"content"`    // 内容
 	Slug      string `json:"slug"`       // 别名
 	Status    string `json:"status"`     // 状态：draft, published
-	AuthorID  int64  `json:"author_id"`  // 作者ID
+	AuthorID  uint   `json:"author_id"`  // 作者ID
 	CreatedAt string `json:"created_at"` // 创建时间
 	UpdatedAt string `json:"updated_at"` // 更新时间
 }
@@ -40,7 +40,7 @@ type PostCreate struct {
 	Content  string `json:"content" validate:"required"`
 	Slug     string `json:"slug" validate:"required"`
 	Status   string `json:"status" validate:"required,oneof=draft published"`
-	AuthorID int64  `json:"author_id" validate:"required"`
+	AuthorID uint   `json:"author_id" validate:"required"`
 }
 
 // PostUpdate represents the data needed to update an existing post
@@ -49,5 +49,5 @@ type PostUpdate struct {
 	Content  string `json:"content" validate:"required"`
 	Slug     string `json:"slug" validate:"required"`
 	Status   string `json:"status" validate:"required,oneof=draft published"`
-	AuthorID int64  `json:"author_id" validate:"required"`
+	AuthorID uint   `json:"author_id" validate:"required"`
 }
